requirements.txt
python-telegram-bot==20.7
requests
reportlab
apscheduler
python-dotenv
